+++ mods/mobs_redo/mobs_ghost_redo/init.lua
@@ -1,44 +1,10 @@
---[[
-	Mobs Ghost Redo - Adds ghosts.
-	Copyright © 2018, 2019 Hamlet <hamlatmesehub@riseup.net> and contributors.
-
-	Licensed under the EUPL, Version 1.2 or – as soon they will be
-	approved by the European Commission – subsequent versions of the
-	EUPL (the "Licence");
-	You may not use this work except in compliance with the Licence.
-	You may obtain a copy of the Licence at:
-
-	https://joinup.ec.europa.eu/software/page/eupl
-	https://eur-lex.europa.eu/legal-content/EN/TXT/?uri=CELEX:32017D0863
-
-	Unless required by applicable law or agreed to in writing,
-	software distributed under the Licence is distributed on an
-	"AS IS" basis,
-	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
-	implied.
-	See the Licence for the specific language governing permissions
-	and limitations under the Licence.
-
---]]
 
 -- Translation and settings
 
 local S = core.get_translator("mobs_ghost_redo")
-local ghost_daytime_check = core.settings:get_bool("mobs_ghost_redo_daytime_check")
-local ghost_bones_only = core.settings:get_bool("mobs_ghost_redo_bones_only")
-local ghost_difficulty = core.settings:get_bool("mobs_ghost_redo_difficulty")
-
-if ghost_daytime_check == nil then
-	ghost_daytime_check = false
-end
-
-if ghost_bones_only == nil then
-	ghost_bones_only = false
-end
-
-if ghost_difficulty == nil then
-	ghost_difficulty = false
-end
+local daytime_check = core.settings:get_bool("mobs_ghost_redo_daytime_check")
+local bones_only = core.settings:get_bool("mobs_ghost_redo_bones_only")
+local difficulty = core.settings:get_bool("mobs_ghost_redo_difficulty")
 
 -- Mineclone check
 
@@ -46,65 +12,66 @@ local mod_mcl = core.get_modpath("mcl_core")
 
 -- Spawn settings (default, bones only, mineclone)
 
-local SPAWNING_CHANCE = 7500
-local ACTIVE_OBJECTS = 2
-local SPAWNING_NODES = {"group:cracky", "group:crumbly"}
+local spawn_chance = 7500
+local ghosts_active = 2
+local spawn_nodes = {"group:cracky", "group:crumbly"}
 
-if mod_mcl == true then
-	SPAWNING_NODES = {"group:pickaxey", "group:shovely"}
+if mod_mcl then
+	spawn_nodes = {"group:pickaxey", "group:shovely"}
 end
 
-if ghost_bones_only == true then
-	SPAWNING_NODES = {"bones:bones", "mobs_humans:human_bones"}
-	SPAWNING_CHANCE = 7
-	ACTIVE_OBJECTS = 1
+if bones_only then
+	spawn_nodes = {"bones:bones", "mobs_humans:human_bones"}
+	spawn_chance = 7
+	ghosts_active = 1
 end
 
 -- Functions
 
 local function is_daytime()
-
-	local time = core.get_timeofday() * 24000
-
-	if time >= 4700 and time <= 19250 then return true end
-end
-
-local function random_mesh()
-
-	if math.random(2) == 2 then
-		return "mobs_ghost_redo_ghost_2.b3d"
-	end
-
-	return "mobs_ghost_redo_ghost_1.b3d"
+	local time = core.get_timeofday() ; return time > 0.19 and time < 0.8
 end
 
 -- Drops
 
-local DROPS = {{name = (mod_mcl and "mcl_raw_ores:raw_gold" or "default:gold_lump"),
-		chance = 100, min = 1, max = 5}}
+local mob_drops = {{
+	name = (mod_mcl and "mcl_raw_ores:raw_gold" or "default:gold_lump"),
+	chance = 100, min = 1, max = 5
+}}
 
 -- Entity definition
 
 mobs:register_mob("mobs_ghost_redo:ghost", {
+	description = S("Ghost"),
 	type = "monster",
+	attack_type = "dogfight",
+	group_attack = true,
+	attack_animals = true,
+	reach = 4,
+	damage = 4,
+	water_damage = 0,
+	lava_damage = 0,
+	fire_damage = 0,
+	light_damage = 2,
+	suffocation = false,
+	blood_amount = 0,
 	hp_min = 10,
 	hp_max = 20,
 	armor = 100,
+	collisionbox = {-0.3, -0.5, -0.3, 0.3, 1.5, 0.3},
+	visual = "mesh",
+	mesh = "mobs_ghost_redo_ghost_1.b3d",
+	visual_size = {x = 1, y = 1},
+	textures = {
+		{"mobs_ghost_redo_ghost.png"},
+		{"mobs_ghost_redo_ghost2.png"}
+	},
+	glow = 5,
 	walk_velocity = 1,
 	run_velocity = 4,
 	walk_chance = 25,
 	fly = true,
 	view_range = 15,
-	reach = 4,
-	damage = 4,
-	water_damage = 0,
-	lava_damage = 0,
-	light_damage = 2,
-	suffocation = false,
-	attack_animals = true,
-	group_attack = true,
-	attack_type = "dogfight",
-	blood_amount = 0,
 	makes_footstep_sound = false,
 	sounds = {
 		random = "mobs_ghost_redo_ghost_1",
@@ -113,12 +80,7 @@ mobs:register_mob("mobs_ghost_redo:ghost", {
 		damage = "mobs_ghost_redo_ghost_hit",
 		death = "mobs_ghost_redo_ghost_death"
 	},
-	drops = DROPS,
-	visual = "mesh",
-	visual_size = {x = 1, y = 1},
-	collisionbox = {-0.3, -0.5, -0.3, 0.3, 1.5, 0.3},
-	textures = {"mobs_ghost_redo_ghost.png"},
-	mesh = "mobs_ghost_redo_ghost_1.b3d",
+	drops = mob_drops,
 	animation = {
 		stand_start = 0, stand_end = 80, stand_speed = 15,
 		walk_start = 102, walk_end = 122, walk_speed = 12,
@@ -130,9 +92,7 @@ mobs:register_mob("mobs_ghost_redo:ghost", {
 
 	on_spawn = function(self, pos)
 
-		if ghost_difficulty == true then
-
-			self.health = math.random(20, 30)
+		if difficulty then
 
 			self.immune_to = {
 				{(mod_mcl and "mcl_tools:sword_iron" or "default:sword_steel"), 6},
@@ -147,15 +107,12 @@ mobs:register_mob("mobs_ghost_redo:ghost", {
 			}
 		end
 
-		self.spawned = true
-		self.mesh = random_mesh()
-		self.counter = 0
+		self.mesh = "mobs_ghost_redo_ghost_" .. math.random(2) .. ".b3d"
+		self.light_damage = daytime_check and 0 or 2
+
 		self.object:set_properties({
-			health = self.health,
 			immune_to = self.immune_to,
-			spawned = self.spawned,
 			mesh = self.mesh,
-			counter = self.counter,
 			physical = false,
 			collide_with_objects = false
 		})
@@ -165,57 +122,14 @@ mobs:register_mob("mobs_ghost_redo:ghost", {
 
 	do_custom = function(self, dtime)
 
-		if ghost_daytime_check == true then
+		-- 5 second counter
+		self.counter = (self.counter or 0) + dtime
+		if self.counter < 5 then return end
+		self.counter = 0
 
-			if self.light_damage ~= 0 then
-
-				self.light_damage = 0
-
-				self.object:set_properties({
-					light_damage = self.light_damage
-				})
-			end
-
-			if self.spawned == true then
-
-				if is_daytime() == true then
-					self.object:remove()
-				else
-					self.spawned = false
-					self.object:set_properties({
-						spawned = self.spawned
-					})
-				end
-			else
-				if self.counter < 15.0 then
-
-					self.counter = self.counter + dtime
-
-					self.object:set_properties({
-						counter = self.counter
-					})
-				else
-
-					if is_daytime() == true then
-						self.object:remove()
-					else
-						self.counter = 0
-
-						self.object:set_properties({
-							counter = self.counter
-						})
-					end
-				end
-			end
-		else
-			if self.light_damage ~= 2 then
-
-				self.light_damage = 2
-
-				self.object:set_properties({
-					light_damage = self.light_damage
-				})
-			end
+		-- when setting enabled and daytime remove ghost
+		if daytime_check and is_daytime() then
+			self.object:remove() ; return false
 		end
 	end
 })
@@ -229,25 +143,28 @@ if input then
 	input:close() ; input = nil ; dofile(MP .. "spawn.lua")
 else
 	mobs:spawn({name = "mobs_ghost_redo:ghost",
-		nodes = SPAWNING_NODES,
+		nodes = spawn_nodes,
 		neighbors = {"air"},
 		max_light = 4,
 		interval = 60,
-		chance = SPAWNING_CHANCE,
-		active_object_count = ACTIVE_OBJECTS,
-		min_height = -30912,
+		chance = spawn_chance,
+		active_object_count = ghosts_active,
 		day_toggle = false
 	})
 end
 
--- Ghost's egg
+-- Spawn egg
 
-mobs:register_egg("mobs_ghost_redo:ghost", S("Ghost Spawner"),
-	"mobs_ghost_redo_egg_ghost.png", 0)
+mobs:register_egg("mobs_ghost_redo:ghost", S("Ghost"), "mobs_ghost_redo_egg_ghost.png", 0)
 
 -- Alias
 
 mobs:alias_mob("mobs:ghost", "mobs_ghost_redo:ghost")
 
+-- Lucky Blocks
+
+if core.get_modpath("lucky_block") then
+	lucky_block:add_blocks({ {"spw", "mobs_ghost_redo:ghost", 3} })
+end
 
 print("[MOD] Mobs Ghost Redo loaded.")