+++ mods/mobs_redo/mobs_ghost_redo/spawn_example.lua
@@ -8,6 +8,5 @@ mobs:spawn({name = "mobs_ghost_redo:ghost",
 	interval = 60,
 	chance = 7500,
 	active_object_count = 2,
-	min_height = -30912,
 	day_toggle = false
 })